♻️ Refactor: separate social edit and chat edit
소셜 프로필과 대화 프로필 별도의 컴포넌트로 분리
